refactor: remove createImageUseCase from create task , adjust other files to avoid errors
