+++ src/infrastructure/routes/task.routes.ts
@@ -1,23 +1,13 @@
 import { Router, Request, Response, NextFunction } from "express";
 import { TaskController } from "../controllers/task.controller";
-import {
-  GetTaskUseCase,
-  CreateTaskUseCase,
-  ProcessImageUseCase,
-} from "../../application/use-cases";
-import { TaskRepository } from "../../domain/task.repository";
+import { GetTaskUseCase, CreateTaskUseCase } from "../../application/use-cases";
+import { TaskRepository } from "../../domain/repositories/task.repository";
 
-export default function taskRoutes(
-  taskRepository: TaskRepository,
-  processImageUseCase: ProcessImageUseCase
-) {
+export default function taskRoutes(taskRepository: TaskRepository) {
   const router = Router();
 
   const getTaskUseCase = new GetTaskUseCase(taskRepository);
-  const createTaskUseCase = new CreateTaskUseCase(
-    taskRepository,
-    processImageUseCase
-  );
+  const createTaskUseCase = new CreateTaskUseCase(taskRepository);
   const taskController = new TaskController(getTaskUseCase, createTaskUseCase);
 
   router.get(
